Remove top-level symbols. Move more stuff
