+++ truss/templates/control/control/helpers/truss_patch/model_code_patch_applier.py
@@ -2,6 +2,7 @@ import logging
 import os
 from pathlib import Path
 
+# TODO(marius): remove try-except after TaT.
 # TODO(pankaj) In desparate need of refactoring into separate library
 try:
     from helpers.custom_types import Action, Patch
