+++ truss/templates/control/control/helpers/truss_patch/model_container_patch_applier.py
@@ -16,7 +16,17 @@ from helpers.custom_types import (
 )
 from helpers.errors import UnsupportedPatch
 from helpers.truss_patch.model_code_patch_applier import apply_code_patch
-from truss.base.truss_config import ExternalData, ExternalDataItem, TrussConfig
+
+# TODO(marius): remove try-except after TaT.
+try:
+    from truss.base.truss_config import ExternalData, ExternalDataItem, TrussConfig
+except ImportError:
+    from truss.truss_config import (  # type: ignore[no-redef]
+        ExternalData,
+        ExternalDataItem,
+        TrussConfig,
+    )
+
 from truss.util.download import download_external_data
